Switched to tokio and reqwest
Originally intended on using smol and surf to keep the runtime and
dependencies small to reduce time to compile and make a smaller binary.

The switch was made to make it easier to contribute as tokio is the most
popular async runtime and reqwest is the most popular high level HTTP Client library.

Hypothetically this change could improve performance but I expect most
of the work be being network/io